refactor: implemented the ChecksumProvider interface
(cherry picked from commit 5a5709fafd492bbc9669b3d24c2c447a51627d23)
